AVX-32997: Refactor modules
