+++ common/constants.py
@@ -1,0 +1,3 @@
+HANDLE_HA_TIMEOUT = 840  # 14 min
+WAIT_DELAY = 30
+INITIAL_SETUP_DELAY = 10
