+++ csp/instance.py
@@ -1,4 +1,5 @@
 """ CSP APis related to instances"""
+import botocore
 
 
 def get_controller_instance(ec2_client, instance_name, inst_id):
@@ -26,3 +27,15 @@ def get_controller_instance(ec2_client, instance_name, inst_id):
             instance_name, inst_id_err, str(err))
         print(describe_err)
     return describe_err, controller_instanceobj
+
+
+def enable_t2_unlimited(client, inst_id):
+    """ Modify instance credit to unlimited for T2 """
+    print("Enabling T2 unlimited for %s" % inst_id)
+    try:
+        client.modify_instance_credit_specification(ClientToken=inst_id,
+                                                    InstanceCreditSpecifications=[{
+                                                        'InstanceId': inst_id,
+                                                        'CpuCredits': 'unlimited'}])
+    except botocore.exceptions.ClientError as err:
+        print(str(err))
