+++ tools/string_utils.py
